+++ bot/handlers/__init__.py
@@ -1,5 +1,6 @@
 from aiogram import Router
 
+from .accounting_documents import router as accounting_documents_router
 from .accounting_document_intake import router as accounting_document_intake_router
 from .contacts import router as contacts_router
 from .contracts import router as contracts_router
@@ -17,6 +18,7 @@ routers: list[Router] = [
     onboarding_router,
     supplier_router,
     accounting_document_intake_router,
+    accounting_documents_router,
     officeflow_attachment_router,
     contacts_router,
     contracts_router,